Read property when new_value is set, don't rely on SelectionNotify
Not all SelectionOwners will send a SelectionNotify but they will
update the target property. So handle that by resetting request_active
when reading (and deleting) the property but still allow for
SelectionNotify to be received later. Use the timestamp of the
SelectionNotify to keep track of which request it is a reply to.